requirements.txt
frida
frida-tools
jmespath
colorama
pandas
numpy
pefile
pycryptodome
pyOpenSSL
py7zr